% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asem.R
\name{asem}
\alias{asem}
\title{asem()}
\usage{
asem(
  Loo,
  K,
  t0 = -0.5,
  aW = 0.01,
  bW = 3.04,
  Lm,
  dLm = NULL,
  tmin = 0,
  tmax,
  h,
  M,
  Lc = NULL,
  dLc = NULL,
  Fvec = NULL,
  R0 = 1
)
}
\arguments{
\item{Loo}{infinitive length (cm)}

\item{K}{brody  growth coefficient K}

\item{t0}{theoretical age a length zero (default -0.5)}

\item{aW}{parameter of the lenght-weight relation (default 0.01)}

\item{bW}{parameter of the lenght-weight relation (default 3.04)}

\item{Lm}{length-at-50-maturity (cm)}

\item{dLm}{width of ogive (default=0.1*Lm)}

\item{tmin}{minimum age}

\item{tmax}{maximum age}

\item{h}{steepness of B&H SSR}

\item{M}{natural mortality}

\item{Lc}{length at first capture}

\item{dLc}{width of ogive (default = 0.1*Lc)}

\item{Fvec}{vector of instantanious fishing fortality}

\item{R0}{unfished recruitment default = 1}
}
\value{
spm pars, equilibrium dynamics, life history table
}
\description{
Age-Structured Equilibrium Model to compute ratio Fmsy= MSY/Bmsy and Bmsy/K
}
\author{
Henning Winker (JRC-EC)
}
