% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl2asem.R
\name{fl2asem}
\alias{fl2asem}
\title{fl2asem()}
\usage{
fl2asem(
  flstk = flstk,
  mc = 1000,
  t0 = -0.5,
  aW = 0.01,
  bW = 3.04,
  dLm = NULL,
  tmin = 0,
  Lc = NULL,
  dLc = NULL,
  Fvec = NULL,
  R0 = 1,
  plot.progress = FALSE
)
}
\arguments{
\item{flstk}{output object from flmvn_traits()}

\item{mc}{number of Monte-Carlo replicates extraced from MVN draws}

\item{t0}{theoretical age a length zero (default -0.5)}

\item{aW}{parameter of the lenght-weight relation (default 0.01)}

\item{bW}{parameter of the lenght-weight relation (default 3.04)}

\item{dLm}{width of ogive (default=0.1*Lm)}

\item{tmin}{minimum age (default = 0)}

\item{Lc}{length at first capture (default = Lm)}

\item{dLc}{width of ogive (default = 0.1*Lc)}

\item{Fvec}{vector of instantanious fishing fortality}

\item{R0}{unfished recruitment default = 1}

\item{plot.progress}{plot simulated production function}
}
\value{
Pella-Tomlison Prior
}
\description{
Translates MVN stocks pars from FishLife (Thorson et al. 2020) 
into Pella-Tomlison SPM priors r, shape, bmsy/k (JABBA/SPicT type) 
through an Age-Structured Equilibrium Model (Winker et al. 2020)
}
\author{
Henning Winker (JRC-EC)
}
