% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flmvn_traits.R
\name{flmvn_traits}
\alias{flmvn_traits}
\title{flmvn_traits()}
\usage{
flmvn_traits(
  Genus = "Rhabdosargus",
  Species = "globiceps",
  Loo = NULL,
  K = NULL,
  tmax = NULL,
  tm = NULL,
  M = NULL,
  Lm = NULL,
  h = NULL,
  nmc = 2 * 10^5,
  upper.quant = 0.9,
  Plot = TRUE,
  savepng = FALSE,
  PlotPath = getwd()
)
}
\arguments{
\item{Genus}{genus name must match FishBase}

\item{Species}{species name must match FishBase}

\item{Loo}{optional tuning prior dist for Linf c(mu, cv)}

\item{K}{optional tuning prior dist for brody K c(mu, cv)}

\item{tmax}{optional tuning prior dist for maximum age c(mu, cv)}

\item{tm}{optional tuning prior dist for age maturity c(mu, cv)}

\item{Lm}{optional tuning prior dist for length maturity c(mu, cv)}

\item{h}{optional uniform tuning prior dist for steepness h c(min, max)}

\item{nmc}{number Monte-Carlo MVN draws}

\item{upper.quant}{upper quatile for CIs ,}

\item{Plot}{option TRUE/FALSE to switch of plots}

\item{savepng}{option TRUE/FALSE to save plots as png}

\item{PlotPath}{directory to save plot if savepng=TRUE}
}
\value{
stock parameter list, mvn covar, Monte-Carlo matrix of stock pars
}
\description{
Generate stock parameters with tuning from FishLife (Thorson et al. 2000) using MVN
}
\author{
Henning Winker (JRC-EC)
}
