% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorutils.R
\name{from_logith}
\alias{from_logith}
\title{from_logith}
\usage{
from_logith(logit_h)
}
\arguments{
\item{logit_h}{logit converted steepness h}
}
\value{
steepness h
}
\description{
convert steepness from logit
}
