% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leslie_r.R
\name{leslie_r}
\alias{leslie_r}
\title{leslie_r()}
\usage{
leslie_r(
  Loo = 80,
  K = 0.2,
  t0 = -0.5,
  aW = 0.01,
  bW = 3.04,
  mat = c(35, 0),
  dmat = NULL,
  minage = 0,
  maxage = 12,
  M = 0.25,
  h = 0.7
)
}
\arguments{
\item{Loo}{infinitive length (cm)}

\item{K}{brody  growth coefficient K}

\item{t0}{theoretical age a length zero (default -0.5)}

\item{aW}{parameter of the lenght-weight relation (default 0.01)}

\item{bW}{parameter of the lenght-weight relation (default 3.04)}

\item{mat}{length- c(Length,0) or age-at-50-maturity c(Age,1) (default is length, i.e. c(Lenght,0))}

\item{dmat}{width of ogive (default=0.1*Mat)}

\item{M}{natural mortality}

\item{h}{steepness of BevHolt SSR}

\item{tmin}{minimum age}

\item{tmax}{maximum age}
}
\value{
r, G (generation time)
}
\description{
Generate r from Leslie Matrix assuming BevHolt SSR (Afte MacAllister)
}
\author{
Henning Winker (JRC-EC)
}
