% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorutils.R
\name{to_logith}
\alias{to_logith}
\title{to_logith}
\usage{
to_logith(h)
}
\arguments{
\item{logit_h}{logit converted steepness h}
}
\value{
logit transformed steepness h
}
\description{
convert steepness to logit
}
